Move HitObjectComposerTestCase to osu.Game assembly
